Спринт 16.1.Settings.3 переходим на идею с фичами. main уже на месте.
