+++ app/src/main/java/com/practicum/playlistmaker/domain/sharing/impl/SharingInteractorImpl.kt
@@ -1,32 +1,32 @@
 package com.practicum.playlistmaker.domain.sharing.impl
 
 import com.practicum.playlistmaker.domain.sharing.SharingInteractor
-import com.practicum.playlistmaker.data.sharing.ExternalNavigator
+import com.practicum.playlistmaker.domain.ExternalNavigator
 
 class SharingInteractorImpl(
     private val externalNavigator: ExternalNavigator,
 ) : SharingInteractor {
     override fun shareApp() {
-        externalNavigator.shareLink(getShareAppLink())
+//        externalNavigator.shareLink(getShareAppLink())
     }
 
     override fun openTerms() {
-        externalNavigator.openLink(getTermsLink())
+//        externalNavigator.openLink(getTermsLink())
     }
 
     override fun openSupport() {
-        externalNavigator.openEmail(getSupportEmailData())
+//        externalNavigator.openEmail(getSupportEmailData())
     }
 
-    private fun getShareAppLink(): String {
-        // Нужно реализовать
-    }
-
-    private fun getSupportEmailData(): EmailData {
-        // Нужно реализовать
-    }
-
-    private fun getTermsLink(): String {
-        // Нужно реализовать
-    }
+//    private fun getShareAppLink(): String {
+//        // Нужно реализовать
+//    }
+//
+//    private fun getSupportEmailData(): EmailData {
+//        // Нужно реализовать
+//    }
+//
+//    private fun getTermsLink(): String {
+//        // Нужно реализовать
+//    }
 }
