+++ todo.removeafter/context_refactoring.md
@@ -1,0 +1,135 @@
+# Контекст рефакторинга PlaylistMaker на MVVM
+
+## Дата: 2025-12-26
+
+## Что уже сделано ✅
+
+1. **CLAUDE.md обновлён** — добавлено правило №5 о запрете готового кода
+2. **Main экран рефакторен:**
+   - Перенесён в `com.practicum.playlistmaker.main.ui.MainActivity`
+   - Добавлен ViewBinding
+   - Упрощён код (убраны findViewById, оптимизированы Intent)
+   - AndroidManifest автоматически обновился
+   - ViewModel не создавалась (там только навигация, нет логики)
+3. **Закоммичены** начальные изменения в Settings (интерфейсы SettingsRepository и SettingsInteractor)
+
+## Текущий статус TODO:
+
+- [✅] Закоммитить текущие незавершенные изменения в Settings
+- [✅] Рефакторинг Main на MVVM (по фичам)
+- [🔄] Рефакторинг Settings на MVVM (по фичам) **← ТЕКУЩАЯ ЗАДАЧА**
+- [ ] Рефакторинг Sharing на MVVM (по фичам)
+- [ ] Рефакторинг Search на MVVM (по фичам)
+- [ ] Рефакторинг Player на MVVM (по фичам)
+- [ ] Рефакторинг MediaLibrary на MVVM (по фичам)
+- [ ] Проверка работы всего приложения после рефакторинга
+
+## Целевая структура проекта (по фичам):
+
+```
+com.practicum.playlistmaker/
+├── main/
+│   └── ui/
+│       └── MainActivity.kt ✅ ГОТОВО
+├── settings/
+│   ├── data/
+│   ├── domain/
+│   └── ui/
+├── sharing/
+│   ├── data/
+│   ├── domain/
+│   └── ui/
+├── search/
+│   ├── data/
+│   ├── domain/
+│   └── ui/
+├── player/
+│   ├── data/
+│   ├── domain/
+│   └── ui/
+└── medialibrary/
+    ├── data/
+    ├── domain/
+    └── ui/
+```
+
+## Проблемы обнаруженные в Settings:
+
+### Файлы Settings (дубли и незавершённые):
+
+**UI слой:**
+- `ui/settings/activity/SettingsActivity.kt` — актуальный
+- `ui/settings/view_model/SettingsViewModel.kt` — актуальный (использует domain.settings.SettingsInteractor)
+
+**Domain слой (ДУБЛИ!):**
+- `domain/SettingsRepository.kt` — **НОВЫЙ** интерфейс (методы: getThemeSettings(), updateThemeSetting() с моделью ThemeSettings)
+- `domain/api/SettingsRepository.kt` — **СТАРЫЙ** интерфейс (методы: getDarkTheme(), saveDarkTheme() с Boolean)
+- `domain/settings/SettingsInteractor.kt` — интерфейс (новый, с ThemeSettings)
+- `domain/settings/impl/SettingsInteractorImpl.kt` — **ПУСТОЙ КЛАСС** (не доделано)
+- `domain/settings/model/ThemeSettings.kt` — модель
+
+**Data слой (ДУБЛИ!):**
+- `data/repository/SettingsRepositoryImpl.kt` — старая реализация (используется в Creator!)
+- `data/settings/impl/SettingsRepositoryImpl.kt` — новая реализация?
+
+### Что использует Creator (зависимости):
+```
+Creator.kt использует:
+- domain.api.SettingsRepository (СТАРЫЙ интерфейс)
+- data.repository.SettingsRepositoryImpl (СТАРАЯ реализация)
+```
+
+## План рефакторинга Settings (пошаговый):
+
+**Шаг 1: Инвентаризация**
+- Изучить все файлы Settings
+- Понять какие актуальные, какие дубли
+- Определить что используется в коде
+
+**Шаг 2: Создать структуру папок**
+- Создать `settings/data/`, `settings/domain/`, `settings/ui/`
+- Ничего не переносить, просто создать пустые папки
+- ✅ Проверка компиляции
+
+**Шаг 3: Перенос UI-слоя**
+- Перенести SettingsActivity в `settings/ui/activity/`
+- Перенести SettingsViewModel в `settings/ui/view_model/`
+- Обновить импорты
+- Обновить AndroidManifest если нужно
+- ✅ Проверка компиляции
+
+**Шаг 4: Перенос Domain-слоя**
+- Перенести актуальные интерфейсы в `settings/domain/`
+- Перенести модели в `settings/domain/model/`
+- Перенести impl в `settings/domain/impl/`
+- Обновить импорты
+- ✅ Проверка компиляции
+
+**Шаг 5: Перенос Data-слоя**
+- Перенести Repository impl в `settings/data/`
+- Обновить импорты
+- Обновить Creator
+- ✅ Проверка компиляции
+
+**Шаг 6: Удаление дублей**
+- Удалить старые файлы
+- ✅ Проверка компиляции
+
+## Важные замечания:
+
+1. **Проблема с импортами** — после незавершённого рефакторинга много файлов с неправильными импортами. Пример: SharingInteractorImpl импортировал ExternalNavigator из data вместо domain.
+
+2. **Чистая архитектура:**
+   - Domain НЕ зависит от Data
+   - Domain использует только интерфейсы
+   - Data реализует интерфейсы из Domain
+
+3. **После каждого шага** делать Build (Ctrl+F9) и проверять что нет ошибок компиляции
+
+4. **Коммитить** после каждого успешно завершённого экрана
+
+## Следующие действия:
+
+1. Начать Шаг 1 для Settings — инвентаризация файлов
+2. Определить какие файлы актуальные
+3. Продолжить по плану
